refactor: moved loginUtil to composables folder
